requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.10.0
    # via nbconvert
bleach==4.1.0
    # via nbconvert
cffi==1.15.0
    # via argon2-cffi-bindings
colorama==0.4.4
    # via ipython
cycler==0.11.0
    # via matplotlib
debugpy==1.5.1
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fonttools==4.31.1
    # via matplotlib
ipykernel==6.9.2
    # via notebook
ipython==8.1.1
    # via
    #   -r requirements.in
    #   ipykernel
ipython-genutils==0.2.0
    # via notebook
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   nbconvert
    #   notebook
jsonschema==4.4.0
    # via nbformat
jupyter-client==7.1.2
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.2
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
kiwisolver==1.4.0
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.1
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclient==0.5.13
    # via nbconvert
nbconvert==6.4.4
    # via notebook
nbformat==5.2.0
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.10
    # via -r requirements.in
numpy==1.22.3
    # via matplotlib
packaging==21.3
    # via
    #   bleach
    #   matplotlib
    #   pytesseract
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pickleshare==0.7.5
    # via ipython
pillow==9.0.1
    # via
    #   matplotlib
    #   pytesseract
prometheus-client==0.13.1
    # via notebook
prompt-toolkit==3.0.28
    # via ipython
psutil==5.9.0
    # via ipykernel
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.11.2
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
pytesseract==0.3.9
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
pywin32==303
    # via jupyter-core
pywinpty==2.0.5
    # via terminado
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
stack-data==0.2.0
    # via ipython
terminado==0.13.3
    # via notebook
testpath==0.6.0
    # via nbconvert
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach

# The following packages are considered to be unsafe in a requirements file:
# setuptools
